Wifi: Check for SAE/OWE AKM prior to PSK
PSK-SHA-256 matches with the PSK check and tries to
configure PSK AKM to wpa_supplicant, which fails.
This could be acceptable as framework does not
support PSK-SHA-256 but this would still fail when
the AP supports two AKM's (PSK-SHA-256 and SAE/OWE).
STA supporting WPA3 AKM's should atleast try for
SAE/OWE AKM's prior to the check for PSK.

This change does the same.

Change-Id: Ia2d465bd2650572118be91fb6eaf62839377ba24
CRs-Fixed: 2223958
Signed-off-by: DennySPB <dennyspb@gmail.com>